pyproject.toml
[project]
name = "beyond-chatgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit==0.7.700",
    "cohere==4.37",
    "openai==1.3.5",
    "pydantic==2.10.1",
    "python-dotenv==1.0.0",
    "tiktoken>=0.6.0",
]
